requirements.txt
astroid==2.2.5
Django==1.11.20
django-classy-tags==0.8.0
django-cms==3.6.0
django-filer==1.4.4
django-formtools==2.1
django-js-asset==1.2.2
django-mptt==0.9.1
django-polymorphic==2.0.3
django-sekizai==0.10.0
django-treebeard==4.3
djangocms-admin-style==1.2.9
djangocms-attributes-field==1.0.0
djangocms-column==1.9.0
djangocms-file==2.2.0
djangocms-googlemap==1.2.0
djangocms-link==2.3.1
djangocms-picture==2.0.8
djangocms-snippet==2.1.0
djangocms-style==2.1.0
djangocms-text-ckeditor==3.7.0
djangocms-video==2.0.5
djangorestframework==3.9.2
easy-thumbnails==2.6
html5lib==1.0.1
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
pylint==2.3.1
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
Unidecode==1.0.23
webencodings==0.5.1
wrapt==1.11.1
